+++ plugin/src/main/java/io/github/brody0125/springtestisolation/gradle/infrastructure/InfrastructureProviders.java
@@ -10,6 +10,7 @@ public final class InfrastructureProviders {
     static {
         registerJdbc(new PostgreSqlJdbcInfrastructureProvider());
         registerJdbc(new MySqlJdbcInfrastructureProvider());
+        registerJdbc(new OracleJdbcInfrastructureProvider());
         registerCache(new RedisCacheInfrastructureProvider());
     }